| Add counter screen as a sample scenario
