+++ sample/src/main/java/com/trendyol/transmission/components/features/colorpicker/ColorPickerSignal.kt
@@ -1,0 +1,8 @@
+package com.trendyol.transmission.components.features.colorpicker
+
+import androidx.compose.ui.graphics.Color
+import com.trendyol.transmission.Transmission
+
+sealed interface ColorPickerSignal : Transmission.Signal {
+    data class SelectColor(val index: Int, val selectedColor: Color) : ColorPickerSignal
+}
