+++ sample/src/main/java/com/trendyol/transmission/components/features/input/WrittenInput.kt
@@ -1,0 +1,5 @@
+package com.trendyol.transmission.components.features.input
+
+import com.trendyol.transmission.Transmission
+
+data class WrittenInput(val value: String) : Transmission.Data
