+++ sample/src/main/java/com/trendyol/transmission/components/features/multioutput/MultiOutputTransformer.kt
@@ -1,9 +1,9 @@
-package com.trendyol.transmission.features.multioutput
+package com.trendyol.transmission.components.features.multioutput
 
 import com.trendyol.transmission.DefaultDispatcher
-import com.trendyol.transmission.features.colorpicker.ColorPickerEffect
-import com.trendyol.transmission.features.input.InputEffect
-import com.trendyol.transmission.features.output.OutputTransformer
+import com.trendyol.transmission.components.features.colorpicker.ColorPickerEffect
+import com.trendyol.transmission.components.features.input.InputEffect
+import com.trendyol.transmission.components.features.output.OutputTransformer
 import com.trendyol.transmission.transformer.Transformer
 import com.trendyol.transmission.transformer.dataholder.dataHolder
 import com.trendyol.transmission.transformer.handler.HandlerRegistry
@@ -11,7 +11,7 @@ import com.trendyol.transmission.transformer.handler.handlers
 import com.trendyol.transmission.transformer.handler.onEffect
 import com.trendyol.transmission.transformer.request.Contracts
 import com.trendyol.transmission.transformer.request.identity
-import com.trendyol.transmission.ui.MultiOutputUiState
+import com.trendyol.transmission.components.features.MultiOutputUiState
 import kotlinx.coroutines.CoroutineDispatcher
 import javax.inject.Inject
