+++ sample/src/main/java/com/trendyol/transmission/components/features/output/OutputCalculationResult.kt
@@ -1,0 +1,5 @@
+package com.trendyol.transmission.components.features.output
+
+import com.trendyol.transmission.Transmission
+
+data class OutputCalculationResult(val result: String) : Transmission.Data
