+++ sample/src/main/java/com/trendyol/transmission/counter/CounterData.kt
@@ -1,0 +1,5 @@
+package com.trendyol.transmission.counter
+
+import com.trendyol.transmission.Transmission
+
+data class CounterData(val id: String) : Transmission.Data
