+++ sample/src/main/java/com/trendyol/transmission/counter/CounterSignal.kt
@@ -1,0 +1,7 @@
+package com.trendyol.transmission.counter
+
+import com.trendyol.transmission.Transmission
+
+sealed interface CounterSignal : Transmission.Signal {
+    data object Lookup : CounterSignal
+}
