+++ sample/src/main/java/com/trendyol/transmission/counter/Holder.kt
@@ -1,0 +1,29 @@
+package com.trendyol.transmission.counter
+
+import com.trendyol.transmission.Transmission
+import com.trendyol.transmission.transformer.Transformer
+import com.trendyol.transmission.transformer.dataholder.dataHolder
+import com.trendyol.transmission.transformer.request.Contracts
+import com.trendyol.transmission.transformer.request.computation.ComputationRegistry
+import com.trendyol.transmission.transformer.request.computation.computations
+import com.trendyol.transmission.transformer.request.computation.register
+import com.trendyol.transmission.transformer.request.computationWithArgs
+import com.trendyol.transmission.transformer.request.identity
+
+val lookUpAndReturn = Contracts.computationWithArgs<String, Int>("lookupAndReturn")
+
+class Holder : Transformer(Contracts.identity("master")) {
+
+    data class TestCounter(val value: Int) : Transmission.Data
+
+    val counterData = dataHolder(TestCounter(0))
+
+    override val computations: ComputationRegistry = computations {
+        register(lookUpAndReturn) { id ->
+            counterData.update { it.copy(value = it.value.plus(1)) }
+            val data = counterData.getValue().value
+            println("Master is $data")
+            data
+        }
+    }
+}
