--- sample/src/main/java/com/trendyol/transmission/features/colorpicker/ColorPickerEffect.kt
@@ -1,9 +1,0 @@
-package com.trendyol.transmission.features.colorpicker
-
-import androidx.compose.ui.graphics.Color
-import com.trendyol.transmission.Transmission
-
-sealed interface ColorPickerEffect: Transmission.Effect {
-	data class BackgroundColorUpdate(val color: Color): ColorPickerEffect
-	data class SelectedColorUpdate(val color: Color): ColorPickerEffect
-}
