--- sample/src/main/java/com/trendyol/transmission/features/input/InputEffect.kt
@@ -1,7 +1,0 @@
-package com.trendyol.transmission.features.input
-
-import com.trendyol.transmission.Transmission
-
-sealed interface InputEffect: Transmission.Effect {
-	data class InputUpdate(val value: String): InputEffect
-}
