--- sample/src/main/java/com/trendyol/transmission/features/input/InputSignal.kt
@@ -1,7 +1,0 @@
-package com.trendyol.transmission.features.input
-
-import com.trendyol.transmission.Transmission
-
-sealed interface InputSignal: Transmission.Signal {
-	data class InputUpdate(val value: String): InputSignal
-}
