--- sample/src/main/java/com/trendyol/transmission/features/input/WrittenInput.kt
@@ -1,5 +1,0 @@
-package com.trendyol.transmission.features.input
-
-import com.trendyol.transmission.Transmission
-
-data class WrittenInput(val value: String): Transmission.Data
