--- sample/src/main/java/com/trendyol/transmission/features/output/OutputCalculationResult.kt
@@ -1,5 +1,0 @@
-package com.trendyol.transmission.features.output
-
-import com.trendyol.transmission.Transmission
-
-data class OutputCalculationResult(val result: String): Transmission.Data
