+++ sample/src/test/java/com/trendyol/transmission/features/colorpicker/ColorPickerTransformerTest.kt
@@ -1,8 +1,11 @@
 package com.trendyol.transmission.features.colorpicker
 
 import androidx.compose.ui.graphics.Color
+import com.trendyol.transmission.components.features.colorpicker.ColorPickerEffect
+import com.trendyol.transmission.components.features.colorpicker.ColorPickerSignal
+import com.trendyol.transmission.components.features.colorpicker.ColorPickerTransformer
 import com.trendyol.transmission.transformer.util.TestCoroutineRule
-import com.trendyol.transmission.ui.ColorPickerUiState
+import com.trendyol.transmission.components.features.ColorPickerUiState
 import com.trendyol.transmissiontest.attachToRouter
 import com.trendyol.transmissiontest.test
 import org.junit.Before
