+++ sample/src/test/java/com/trendyol/transmission/features/output/OutputTransformerTest.kt
@@ -1,10 +1,11 @@
 package com.trendyol.transmission.features.output
 
 import com.trendyol.transmission.effect.RouterEffect
-import com.trendyol.transmission.features.input.InputEffect
+import com.trendyol.transmission.components.features.input.InputEffect
+import com.trendyol.transmission.components.features.output.OutputTransformer
 import com.trendyol.transmission.transformer.util.TestCoroutineRule
-import com.trendyol.transmission.ui.ColorPickerUiState
-import com.trendyol.transmission.ui.OutputUiState
+import com.trendyol.transmission.components.features.ColorPickerUiState
+import com.trendyol.transmission.components.features.OutputUiState
 import com.trendyol.transmissiontest.attachToRouter
 import com.trendyol.transmissiontest.test
 import org.junit.Before
